device: fix attestation report size
Currently we do not account for the size of the attestation reports when
moving them from the C struct to the Rust struct. This results in many
zeros being included at the end of the report.

Since the size of each report is provided, use this to copy only the
bytes which contain the report.

Signed-off-by: Tobin Feldman-Fitzthum <tfeldmanfitz@nvidia.com>