device: fix attestation report size (#116)
* device: fix attestation report size

Currently we do not account for the size of the attestation reports when
moving them from the C struct to the Rust struct. This results in many
zeros being included at the end of the report.

Since the size of each report is provided, use this to copy only the
bytes which contain the report.

Signed-off-by: Tobin Feldman-Fitzthum <tfeldmanfitz@nvidia.com>

* device: check data length for cc cert chains

We are ignoring the size of the attestation reports and
indiscriminately exposing the entire buffer.

Instead, truncate the reports based on the report length.

Signed-off-by: Tobin Feldman-Fitzthum <tfeldmanfitz@nvidia.com>

---------

Signed-off-by: Tobin Feldman-Fitzthum <tfeldmanfitz@nvidia.com>